backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
mercadopago>=2.2.0
httpx>=0.27.0
python-multipart>=0.0.9
aiosmtplib>=3.0.0
jinja2>=3.1.0
slowapi>=0.1.9
